price als BigDecimal

+++ src/main/java/de/iwish/iWish/RestExampleController.java
@@ -4,6 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -13,9 +14,9 @@ public class RestExampleController {
     @GetMapping(path = "/wish")
     public ResponseEntity<List<Wish>> getWishes() {
         List<Wish> wishes = new ArrayList<>();
-        wishes.add(new Wish("Fußball", 10.9, LocalDate.of(2024, 4, 1), "www.amazon.de", Wish.Priority.MITTEL));
-        wishes.add(new Wish("Laptop", 999.99, LocalDate.of(2024, 12, 25), "www.ebay.de", Wish.Priority.HOCH));
-        wishes.add(new Wish("Buch", 14.99, LocalDate.of(2024, 6, 15), "www.buch.de", Wish.Priority.NIEDRIG));
+        wishes.add(new Wish("Fußball", BigDecimal.valueOf(10.9), LocalDate.of(2024, 4, 1), "www.amazon.de", Wish.Priority.MITTEL));
+        wishes.add(new Wish("Laptop", BigDecimal.valueOf(999.99), LocalDate.of(2024, 12, 25), "www.ebay.de", Wish.Priority.HOCH));
+        wishes.add(new Wish("Buch", BigDecimal.valueOf(14.99), LocalDate.of(2024, 6, 15), "www.buch.de", Wish.Priority.NIEDRIG));
         return ResponseEntity.ok(wishes);
     }
 }
